requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
openai==1.38.0
sqlalchemy==1.4.25
python-dotenv==1.0.1
PyPDF2==3.0.1
pandas==2.2.0
scikit-learn==1.5.0
